Renamed the tab bar view controller and added a create subjects view controller function
